backend/requirements.txt
fastapi
openai
python-dotenv
azure-cosmos==4.6.0  
papermill
unicorn
pydantic
pydantic-settings